+++ include/circt/Conversion/Passes.h
@@ -36,6 +36,7 @@
 #include "circt/Conversion/MooreToCore.h"
 #include "circt/Conversion/PipelineToHW.h"
 #include "circt/Conversion/SCFToCalyx.h"
+#include "circt/Conversion/SeqToSV.h"
 #include "circt/Conversion/StandardToHandshake.h"
 #include "circt/Conversion/VerifToSV.h"
 #include "mlir/IR/DialectRegistry.h"
