+++ include/circt/Dialect/Seq/SeqPasses.h
@@ -20,19 +20,14 @@ namespace circt {
 namespace seq {
 
 #define GEN_PASS_DECL_LOWERSEQTOSV
-#define GEN_PASS_DECL_LOWERSEQFIRRTLTOSV
 #define GEN_PASS_DECL_EXTERNALIZECLOCKGATE
 #include "circt/Dialect/Seq/SeqPasses.h.inc"
 
 std::unique_ptr<mlir::Pass>
 createSeqLowerToSVPass(std::optional<bool> lowerToAlwaysFF = {});
-std::unique_ptr<mlir::Pass> createLowerSeqFIRRTLInitToSV();
-std::unique_ptr<mlir::Pass>
-createSeqFIRRTLLowerToSVPass(const LowerSeqFIRRTLToSVOptions &options = {});
 std::unique_ptr<mlir::Pass> createLowerSeqHLMemPass();
 std::unique_ptr<mlir::Pass>
 createExternalizeClockGatePass(const ExternalizeClockGateOptions &options = {});
-std::unique_ptr<mlir::Pass> createLowerFirMemPass();
 std::unique_ptr<mlir::Pass> createLowerSeqFIFOPass();
 
 /// Generate the code for registering passes.
