+++ lib/Conversion/CMakeLists.txt
@@ -10,17 +10,18 @@ add_subdirectory(ExportChiselInterface)
 add_subdirectory(ExportVerilog)
 add_subdirectory(FIRRTLToHW)
 add_subdirectory(FSMToSV)
-add_subdirectory(HandshakeToHW)
 add_subdirectory(HandshakeToDC)
+add_subdirectory(HandshakeToHW)
 add_subdirectory(HWArithToHW)
 add_subdirectory(HWToLLHD)
 add_subdirectory(HWToLLVM)
-add_subdirectory(HWToSystemC)
 add_subdirectory(HWToSV)
+add_subdirectory(HWToSystemC)
 add_subdirectory(LLHDToLLVM)
 add_subdirectory(LoopScheduleToCalyx)
 add_subdirectory(MooreToCore)
 add_subdirectory(PipelineToHW)
 add_subdirectory(SCFToCalyx)
+add_subdirectory(SeqToSV)
 add_subdirectory(StandardToHandshake)
 add_subdirectory(VerifToSV)
