+++ lib/Conversion/SeqToSV/CMakeLists.txt
@@ -1,0 +1,16 @@
+add_circt_conversion_library(CIRCTSeqToSV
+  FirMemLowering.cpp
+  FirRegLowering.cpp
+  SeqToSV.cpp
+  SeqFIRRTLInitToSV.cpp
+  LowerFirMem.cpp
+
+  DEPENDS
+  CIRCTConversionPassIncGen
+
+  LINK_COMPONENTS
+  Core
+
+  LINK_LIBS PUBLIC
+  CIRCTSeq
+)
