+++ lib/Dialect/Seq/Transforms/CMakeLists.txt
@@ -1,6 +1,5 @@
 add_circt_dialect_library(CIRCTSeqTransforms
   ExternalizeClockGate.cpp
-  LowerFirMem.cpp
   LowerSeqHLMem.cpp
   LowerSeqToSV.cpp
   LowerSeqFIFO.cpp
