+++ tools/arcilator/CMakeLists.txt
@@ -8,6 +8,7 @@ target_link_libraries(arcilator
   CIRCTArcTransforms
   CIRCTCombToArith
   CIRCTConvertToArcs
+  CIRCTSeqToSV
   CIRCTSeqTransforms
   CIRCTSupport
   CIRCTTransforms
